before class on thursday11/30
