+++ src/components/loginSignUp/loginSignUp.css
@@ -1,5 +1,5 @@
 body {
   width: 50%;
   margin: auto;
-  padding-top: 100px;
+  padding-top: 50px;
 }
